feat: adiciona classe Menu
